refactor(all) : 도메인 패키지 분리
반영된 코드들 수정
