+++ src/main/java/roomescape/dao/ReservationDao.java
@@ -9,8 +9,8 @@ import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import roomescape.dao.mapper.ReservationRowMapper;
-import roomescape.domain.Reservation;
-import roomescape.domain.ReservationDate;
+import roomescape.domain.reservation.Reservation;
+import roomescape.domain.reservation.ReservationDate;
 
 @Repository
 public class ReservationDao {
