+++ src/main/java/roomescape/dao/ThemeDao.java
@@ -11,8 +11,8 @@ import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import roomescape.dao.mapper.ThemeRowMapper;
-import roomescape.domain.Theme;
-import roomescape.domain.VisitDate;
+import roomescape.domain.reservation.Theme;
+import roomescape.domain.reservation.VisitDate;
 
 @Repository
 public class ThemeDao {
