+++ src/main/java/roomescape/dao/mapper/ReservationTimeRowMapper.java
@@ -6,7 +6,7 @@ import java.time.LocalTime;
 
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Component;
-import roomescape.domain.ReservationTime;
+import roomescape.domain.reservation.ReservationTime;
 
 @Component
 public class ReservationTimeRowMapper implements RowMapper<ReservationTime> {
