+++ src/main/java/roomescape/domain/reservation/Name.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.reservation;
 
 public record Name(String name) {
