+++ src/main/java/roomescape/domain/reservation/Reservation.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.reservation;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
