+++ src/main/java/roomescape/domain/reservation/ReservationTime.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.reservation;
 
 import java.time.DateTimeException;
 import java.time.LocalTime;
