+++ src/main/java/roomescape/domain/reservation/Theme.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.reservation;
 
 import java.util.Objects;
