+++ src/main/java/roomescape/domain/reservation/Thumbnail.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.reservation;
 
 import java.util.List;
 import java.util.regex.Pattern;
