+++ src/main/java/roomescape/service/ThemeService.java
@@ -6,8 +6,8 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.dao.ReservationDao;
 import roomescape.dao.ThemeDao;
-import roomescape.domain.Theme;
-import roomescape.domain.VisitDate;
+import roomescape.domain.reservation.Theme;
+import roomescape.domain.reservation.VisitDate;
 import roomescape.exception.ExistReservationException;
 import roomescape.exception.NotExistException;
 import roomescape.service.dto.input.ThemeInput;
