+++ src/main/java/roomescape/service/dto/input/ReservationTimeInput.java
@@ -1,6 +1,6 @@
 package roomescape.service.dto.input;
 
-import roomescape.domain.ReservationTime;
+import roomescape.domain.reservation.ReservationTime;
 
 public record ReservationTimeInput(String startAt) {
