+++ src/test/java/roomescape/domain/reservation/ThemeTest.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.reservation;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
