+++ src/test/java/roomescape/domain/reservation/VisitDateTest.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.reservation;
 
 import static org.assertj.core.api.Assertions.assertThat;
