+++ src/test/java/roomescape/domain/user/EmailTest.java
@@ -1,9 +1,8 @@
-package roomescape.domain;
+package roomescape.domain.user;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import roomescape.domain.user.Email;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
