+++ src/test/java/roomescape/domain/user/NameTest.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.user;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
