+++ src/test/java/roomescape/domain/user/PasswordTest.java
@@ -1,8 +1,7 @@
-package roomescape.domain;
+package roomescape.domain.user;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.domain.user.Password;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
