+++ src/test/java/roomescape/domain/user/UserTest.java
@@ -1,8 +1,7 @@
-package roomescape.domain;
+package roomescape.domain.user;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.domain.user.*;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
